pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "md-to-pdf"
version = "0.2.1"
description = "Markdown → PDF converter with brand-pack support, watermarking, and CJK fonts"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [{ name = "Sophie Lei", email = "sophie.leiyixiao@gmail.com" }]
keywords = ["markdown", "pdf", "reportlab", "cjk"]
classifiers = [
  "Development Status :: 4 - Beta",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dependencies = [
  "cairosvg>=2.7",
  "click>=8.1",
  "httpx>=0.27",
  "markdown-it-py>=3.0",
  "mdit-py-plugins>=0.4",
  "Pillow>=10.0",
  "pikepdf>=8.0",
  "pydantic>=2.5",
  "pygments>=2.17",
  "pypdf>=4.0",
  "pyyaml>=6.0",
  "reportlab>=4.0",
  "structlog>=24.1",
]

[project.urls]
Homepage = "https://github.com/leiminray/md-to-pdf"
Repository = "https://github.com/leiminray/md-to-pdf"
Issues = "https://github.com/leiminray/md-to-pdf/issues"
Changelog = "https://github.com/leiminray/md-to-pdf/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
  "mypy>=1.8",
  "pdfplumber>=0.11",
  "pypdf>=4.0",
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "pytest-cov>=4.1",
  "ruff>=0.3",
  "types-pygments",
  "types-pyyaml",
  "types-reportlab",
]
docs = [
  "mkdocs>=1.5",
  "mkdocs-material>=9.5",
]
mermaid-pure = ["mermaid>=0.7"]

[project.scripts]
md-to-pdf = "mdpdf.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mdpdf"]

[tool.pytest.ini_options]
addopts = "-ra --strict-markers"
testpaths = ["tests/unit", "tests/integration", "tests/golden"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "S", "C4", "SIM"]

[tool.ruff.lint.per-file-ignores]
# Tests are allowed to use `assert` and otherwise relax security rules.
"tests/**/*.py" = ["S"]

[tool.mypy]
strict = true
python_version = "3.10"
files = ["src/mdpdf"]

# pikepdf does not ship stubs and is imported at module top in
# security/watermark_l2.py + security/deterministic.py. cairosvg / mermaid
# also lack stubs but are imported lazily inside function bodies with
# inline `# type: ignore[import-untyped]` comments — those existing
# suppressions stay in place so we don't suppress import-not-found
# project-wide for them.
[[tool.mypy.overrides]]
module = ["pikepdf", "pikepdf.*"]
ignore_missing_imports = true
